# docs/api-spec/openapi.yml
openapi: 3.0.3
info:
  title: Authorization API
  description: 認可サーバーのAPI群です。
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: API開発チーム
    email: satoshi.nagashiba@gmail.com
servers:
  - url: https://apis.authorization.com
    description: 本番環境
  - url: https://apis-stg.authorization.com
    description: ステージング環境
  - url: https://apis-dev.authorization.com
    description: 開発環境
  - url: https://apis.authorization-php.dev/api
    description: ローカルPHP（Laravel）環境
  - url: https://apis.authorization-csharp.dev/api
    description: ローカルC#（ASP.NET Core）環境
  - url: https://apis.authorization-go-gin.dev/api
    description: ローカルGo（Gin）環境
  - url: https://apis.authorization-go-beego.dev/api
    description: ローカルGo（Beego）環境
  - url: https://apis.authorization-go-echo.dev/api
    description: ローカルGo（Echo）環境
  - url: https://apis.authorization-java.dev/api
    description: ローカルJava（Spring Boot）環境
  - url: https://apis.authorization-kotlin.dev/api
    description: ローカルKotlin（Ktor）環境
  - url: https://apis.authorization-python.dev/api
    description: ローカルPython（FastAPI）環境
  - url: https://apis.authorization-rb-hanami.dev/api
    description: ローカルRuby（Hanami）環境
  - url: https://apis.authorization-rb-rails.dev/api
    description: ローカルRuby（Rails）環境
  - url: https://apis.authorization-rust.dev/api
    description: ローカルRust（Axum）環境
  - url: https://apis.authorization-ts.dev/api
    description: ローカルTypeScript（Hono）環境
paths:
...
  /staffs:
    get:
      operationId: listStaffs
      summary: スタッフ一覧
      tags:
        - staff
      security: []
      parameters:
        - name: keyword
          in: query
          required: false
          description: キーワード
          schema:
            type: string
        - name: roles
          in: query
          required: false
          description: 権限（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: statuses
          in: query
          required: false
          description: 状態（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: page
          in: query
          required: false
          description: 現在ページ番号（フロントのページャー算出用）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          required: false
          description: ソート対象カラム
          schema:
            type: string
            enum: [name, role, status, created_at]
        - name: sort_type
          in: query
          required: false
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffList'
        '404':
          description: ユーザーが存在しない場合のエラー
  /staffs/{id}/updateRole:
    patch:
      operationId: updateStaffRole
      summary: 権限を更新します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - version
                - executor_id
              properties:
                role:
                  type: integer
                  format: int64
                  description: 権限種別（1=管理者、2=メンバー）
                version:
                  type: integer
                  format: int64
                  description: 楽観排他ロック用バージョン番号
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '400':
          description: リクエストが不正です
        '404':
          description: スタッフが存在しない場合のエラー
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
        '500':
          description: サーバーエラー
  /staffs/{id}/restore:
    patch:
      operationId: restoreStaff
      summary: スタッフの論理削除を復元します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - executor_id
              properties:
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '404':
          description: スタッフが存在しない場合のエラー
  /staffs/{id}/delete:
    delete:
      operationId: deleteStaff
      summary: スタッフを論理削除します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
                - executor_id
              properties:
                version:
                  type: integer
                  format: int64
                  description: 楽観排他ロック用バージョン番号
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '404':
          description: スタッフが存在しない場合のエラー
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
...
  /notifications:
    get:
      operationId: listNotifications
      summary: 通知一覧（メタデータ含む）
      description: |
        通知の一覧を返す。各アイテムにメタデータを含む。
        cursor ベースのページネーションを採用する。
      tags:
        - notifications
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: 次ページ取得用の不透明カーソル（初回は省略可）
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: 1ページあたりの件数
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          description: 未認証
        '500':
          description: サーバーエラー

    patch:
      operationId: readAllNotifications
      summary: 全通知を既読にする
      description: |
        ログイン中スタッフの全通知を一括で既読にする。
      tags:
        - notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
        '401':
          description: 未認証
        '500':
          description: サーバーエラー
  /notifications/counts:
    get:
      operationId: getNotificationCounts
      summary: 未読・全件などのサマリー
      description: |
        未読件数・全件数など、通知まわりの集計を返す。
        複数種類のカウントを一度に返す。
      tags:
        - notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationCounts'
        '401':
          description: 未認証
        '500':
          description: サーバーエラー
  /notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: 通知 ID

    patch:
      operationId: readNotification
      summary: 通知を既読にする
      description: |
        1件の通知を既読にする。
      tags:
        - notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
        '400':
          description: リクエストが不正
        '401':
          description: 未認証
        '404':
          description: 通知が存在しない
        '500':
          description: サーバーエラー
components:
  schemas:
...
    Staff:
      type: object
      description: スタッフ一覧の1行です。
      required:
        - id
        - name
        - email
        - role
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          description: 表示名
          example: "山田 太郎"
        email:
          type: string
          format: email
          example: "yamada@example.com"
        role:
          type: integer
          description: 権限コード（1=管理者、2=メンバー）
          example: 1
        status:
          type: integer
          description: 状態コード（1=有効、0=無効。deleted_at の有無で算出）
          example: 1
        created_at:
          type: string
          description: 登録日時（Y-m-d H:i 形式）
          example: "2026-01-15 09:30"
        updated_at:
          type: string
          description: 更新日時（Y-m-d H:i 形式）
          example: "2026-01-20 10:00"
        version:
          type: integer
          format: int64
          description: 楽観排他ロック用バージョン番号
          example: 1
    StaffList:
      type: object
      description: スタッフ一覧（ページング付き）
      required:
        - data
        - pager
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Staff'
        pager:
          $ref: '#/components/schemas/Pager'
...
    Notification:
      type: object
      description: 通知1件
      required:
        - id
        - title
        - message_type
        - read
        - created_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        staff_id:
          type: integer
          format: int64
          example: 1
        message_type:
          type: integer
          description: "1=info / 2=warn / 3=ok"
          example: 1
        title:
          type: string
          example: "新しいクライアントが登録されました"
        message:
          type: string
          example: "株式会社サンプル"
        url:
          type: string
          nullable: true
          description: 遷移先の相対URL。nullの場合はリンクなし
          example: "/clients/show?id=21"
        read:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-01-01T10:00:00+09:00"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-01T10:00:00+09:00"
    NotificationListResponse:
      type: object
      description: cursor ページネーション付き一覧
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        next_cursor:
          type: string
          nullable: true
          description: 次ページがなければ null
    NotificationCounts:
      type: object
      description: 複数種類のカウントを一度に返す
      additionalProperties:
        type: integer
        format: int64
      example:
        unread: 3
        total: 42
    Pager:
      type: object
      description: ページャー情報
      required:
        - count
        - limit
        - next
        - previous
        - page
        - nextPage
        - previousPage
        - pageCount
        - first
        - last
        - firstRecordCount
        - lastRecordCount
        - startPage
        - endPage
      properties:
        count:
          type: integer
          description: 総件数
          example: 1000
        limit:
          type: integer
          description: 取得件数
          example: 20
        next:
          type: boolean
          description: 次頁あり
          example: true
        previous:
          type: boolean
          description: 前頁あり
          example: false
        page:
          type: integer
          description: 現在ページ番号
          example: 1
        nextPage:
          type: integer
          description: 次ページ番号
          example: 2
        previousPage:
          type: integer
          description: 前ページ番号
          example: 0
        pageCount:
          type: integer
          description: 総ページ数
          example: 50
        first:
          type: boolean
          description: 先頭ページへの移動可否（page > 1）
          example: false
        last:
          type: boolean
          description: 最終ページへの移動可否（pageCount > page）
          example: true
        firstRecordCount:
          type: integer
          description: 現在ページの先頭レコード番号
          example: 1
        lastRecordCount:
          type: integer
          description: 現在ページの末尾レコード番号
          example: 20
        startPage:
          type: integer
          description: ページャー表示開始ページ
          example: 1
        endPage:
          type: integer
          description: ページャー表示終了ページ
          example: 10
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientTokenAuth:
      type: http
      scheme: bearer
      description: クライアントのアクセストークン認証（JWT ではなく登録時に発行される固定トークン）
...